% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mapping_general.r
\name{mapping}
\alias{mapping}
\title{mapping}
\usage{
mapping(algorithm, X, V, W, vector_norm, chosen_variable)
}
\arguments{
\item{algorithm}{defines the radial method}

\item{X}{is an N by n matrix wose rows contain the n dimensional data samples.}

\item{V}{is an n by m matrix whose rows define the method's axis vectors.}

\item{W}{is an n by n diagonal matrix defining nonnegative weights for each variable}

\item{vector_norm}{is the vector norm associated with adaptable radial axes plots}

\item{chosen_variable}{is the selected attribute for constrained adaptable radial axes plots}
}
\value{
The low-dimensional embeddings are stored in the N by m matrix P.
}
\description{
Computes the mappings of several radial axes methods
}
\examples{
N = 100
n = 5
m = 2
X <- matrix(rnorm(N, n), nrow = 20, ncol = n)
V <- matrix(rnorm(N, m), nrow = n, ncol = m)
W <- diag(1, nrow = n)
mapping('SC', X, V)
mapping('RadViz', X, V)
mapping('SRA', X, V, W)
mapping('Adaptable', X, V, W, 1)
mapping('Adaptable', X, V, W, 'Inf')
mapping('Adaptable exact', X, V, W, 1, 1)
mapping('Adaptable exact', X, V, W, 2, 1)
mapping('Adaptable exact', X, V, W, 'Inf', 1)
mapping('Adaptable ordered', X, V, W, 1, 1)
mapping('Adaptable ordered', X, V, W, 2, 1)
mapping('Adaptable ordered', X, V, W, 'Inf', 1)
}
